requirements.txt
numpy
pandas
matplotlib
spectral
specdal
python-dotenv
pymongo
streamlit